requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3